+++ backend/scheduler-service/src/main/java/pl/niewadzj/schedulerservice/api/currency/records/CurrencyExternalResponse.java
@@ -2,8 +2,10 @@ package pl.niewadzj.schedulerservice.api.currency.records;
 
 import lombok.Builder;
 
+import java.math.BigDecimal;
+
 @Builder
 public record CurrencyExternalResponse(String currency,
                                        String code,
-                                       Float mid) {
+                                       BigDecimal mid) {
 }
